refactor: LottoConstant - constant 패키지 제거, domain 패키지 하위로 이동
