+++ src/main/java/lotto/domain/Lotto.java
@@ -1,8 +1,8 @@
 package lotto.domain;
 
-import static lotto.constant.LottoConstant.LOTTO_NUMBER_LENGTH;
-import static lotto.constant.LottoConstant.LOTTO_NUMBER_MAX_VALUE;
-import static lotto.constant.LottoConstant.LOTTO_NUMBER_MIN_VALUE;
+import static lotto.domain.LottoConstant.LOTTO_NUMBER_LENGTH;
+import static lotto.domain.LottoConstant.LOTTO_NUMBER_MAX_VALUE;
+import static lotto.domain.LottoConstant.LOTTO_NUMBER_MIN_VALUE;
 import static lotto.exception.InputValidator.ERROR_MESSAGE_PREFIX;
 
 import java.util.ArrayList;
