+++ src/main/java/lotto/domain/LottoConstant.java
@@ -1,4 +1,4 @@
-package lotto.constant;
+package lotto.domain;
 
 /**
  * 비즈니스 로직에 중요하며 글로벌하게 쓰이는 상수들만 모아둔 클래스
